+++ src/helpers/__mocks__/events.ts
@@ -1,7 +1,14 @@
 const event = "push";
 
+const isClientRequired = jest.fn()
+  .mockReturnValue(false)
+  .mockName("helpers.isClientRequired");
+
 const isEventSupported = jest.fn()
   .mockReturnValue([event, true])
   .mockName("helpers.isEventSupported");
 
-export default isEventSupported;
+export {
+  isClientRequired,
+  isEventSupported,
+};