+++ src/helpers/__mocks__/filters.ts
@@ -2,4 +2,6 @@ const getFilters = jest.fn()
   .mockResolvedValue([[], null])
   .mockName("helpers.getFilters");
 
-export default getFilters;
+export {
+  getFilters,
+};