+++ src/helpers/__mocks__/index.ts
@@ -4,6 +4,10 @@ const getFilters = jest.fn()
   .mockResolvedValue([[], null])
   .mockName("helpers.getFilters");
 
+const isClientRequired = jest.fn()
+  .mockReturnValue(false)
+  .mockName("helpers.isClientRequired");
+
 const isEventSupported = jest.fn()
   .mockReturnValue([event, true])
   .mockName("helpers.isEventSupported");
@@ -14,6 +18,7 @@ const parseRawSvgoConfig = jest.fn()
 
 export {
   getFilters,
+  isClientRequired,
   isEventSupported,
   parseRawSvgoConfig,
 };