+++ src/helpers/__mocks__/svgo-config.ts
@@ -2,4 +2,6 @@ const parseRawSvgoConfig = jest.fn()
   .mockReturnValue([{ }, null])
   .mockName("helpers.parseRawSvgoConfig");
 
-export default parseRawSvgoConfig;
+export {
+  parseRawSvgoConfig,
+};