+++ test/unit/helpers/index.test.ts
@@ -1,22 +1,27 @@
-import { mocked } from "ts-jest/utils";
-
 jest.mock("../../../src/helpers/events");
 jest.mock("../../../src/helpers/filters");
 jest.mock("../../../src/helpers/svgo-config");
 
-import _isEventSupported from "../../../src/helpers/events";
-import _getFilters from "../../../src/helpers/filters";
-import _parseRawSvgoConfig from "../../../src/helpers/svgo-config";
+import {
+  isClientRequired,
+  isEventSupported,
+} from "../../../src/helpers/events";
+import {
+  getFilters,
+} from "../../../src/helpers/filters";
+import {
+  parseRawSvgoConfig,
+} from "../../../src/helpers/svgo-config";
 
 import * as helpers from "../../../src/helpers/index";
-
-const isEventSupported = mocked(_isEventSupported);
-const getFilters = mocked(_getFilters);
-const parseRawSvgoConfig = mocked(_parseRawSvgoConfig);
 
 describe("helpers/index.ts", () => {
   test("::getFilters", () => {
     expect(helpers.getFilters).toBe(getFilters);
+  });
+
+  test("::isClientRequired", () => {
+    expect(helpers.isClientRequired).toBe(isClientRequired);
   });
 
   test("::isEventSupported", () => {