+++ src/optimize/index.ts
@@ -1,6 +1,6 @@
 import type { FileSystem } from "../file-systems";
-import type { error, OptimizeProjectData } from "../types";
-import type { Optimizer } from "./types";
+import type { error } from "../types";
+import type { OptimizeProjectData, Optimizer } from "./types";
 
 import errors from "../errors";
 import { optimizeAll } from "./optimize";