+++ src/types.d.ts
@@ -1,7 +1,7 @@
-export type error = string | null;
+import type { GitHub as _GitHub } from "@actions/github/lib/utils";
 
-// Type representing the relevant information of `@actions/github.context`.
-export type Context = {
+// Type representing the relevant API of `@actions/github.context`.
+interface Context {
   readonly eventName: string;
   readonly payload: {
     readonly commits?: {
@@ -17,43 +17,49 @@ export type Context = {
   };
 }
 
-// Type representing the relevant information of `@actions/core`.
-export interface Core extends Inputter, Outputter {
+// Type representing the relevant API of `@actions/core`.
+interface Core extends Inputter, Outputter {
   info(message: string): void;
   setFailed(message: string | Error): void;
   warning(message: string | Error): void;
   debug(message: string): void;
 }
 
-// Type representing data about the optimization process.
-export type OptimizeProjectData = {
-  readonly optimizedCount: number;
-  readonly svgCount: number;
-}
+// Type for errors.
+type error = string | null;
 
-// Type representing the data of a file that is being processed by the Action.
-export type FileData = {
-  readonly content: string;
-  readonly originalEncoding: string;
-  readonly path: string;
+// Type representing the relevant API of `@actions/github`.
+interface GitHub {
+  getOctokit(token: string): Octokit;
+  readonly context: Context;
 }
 
 // Type representing an object from which the Action inputs can be obtained.
-interface InputterOptions {
-  readonly required?: boolean;
-}
-
-export interface Inputter {
+interface Inputter {
   getInput(name: string, options: InputterOptions): string;
   getBooleanInput(name: string, options: InputterOptions): boolean;
 }
 
+// Type representing the options for the `Inputter` interface.
+interface InputterOptions {
+  readonly required?: boolean;
+}
+
+// Type representing `@actions/github`'s Octokit.
+type Octokit = InstanceType<typeof _GitHub>;
+
 // Type representing an object to which Action outputs can be outputted.
-export interface Outputter {
+interface Outputter {
   setOutput(name: string, value: string): void;
 }
 
-export interface GitHub {
-  getOctokit(token: string): any;
-  readonly context: Context;
-}
+export {
+  error,
+  Context,
+  Core,
+  GitHub,
+  Inputter,
+  InputterOptions,
+  Octokit,
+  Outputter,
+};