+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Versioning].
 ## [Unreleased]
 
 - Don't require `repo-token` in scheduled and dispatched events. ([#396])
-- Change `svgo-config` default based on `svgo-version` value. ([#400])
+- Change `svgo-config` default based on `svgo-version` value. ([#400], [#402])
 - Add `repo-token` where missing to the documentation. ([#401])
 
 ## [2.0.0]
@@ -261,5 +261,6 @@ Versioning].
 [#396]: https://github.com/ericcornelissen/svgo-action/pull/396
 [#400]: https://github.com/ericcornelissen/svgo-action/pull/400
 [#401]: https://github.com/ericcornelissen/svgo-action/pull/401
+[#402]: https://github.com/ericcornelissen/svgo-action/pull/402
 [64d0e89]: https://github.com/ericcornelissen/svgo-action/commit/64d0e8958d462695b3939588707815182ecc3690
 [8d8f516]: https://github.com/ericcornelissen/svgo-action/commit/8d8f516583b4340f692e2ea80e1855e5a1211bd3